+++ backend/src/documentation/project.ts
@@ -116,21 +116,7 @@
  *         content:
  *           application/json:
  *             schema:
- *               type: object
- *               properties:
- *                 id:
- *                   type: integer
- *                   example: 1
- *                 name:
- *                   type: string
- *                   example: "Gym"
- *                 color:
- *                   type: string
- *                   example: "#FF5733"
- *                 created_at:
- *                   type: string
- *                 updated_at:
- *                   type: string
+ *               $ref: "#/components/schemas/Project"
  *       400:
  *         description: Bad Request
  *         content: